fix(DashboardTable): move dashboard table, update badge color
